+++ gherkin/go/dialects_builtin.go
@@ -30,9 +30,11 @@ var buildinDialects = gherkinDialectMap{
 				"Agtergrond",
 			},
 			scenario: []string{
+				"Voorbeeld",
 				"Situasie",
 			},
 			scenarioOutline: []string{
+				"Voorbeeld Uiteensetting",
 				"Situasie Uiteensetting",
 			},
 			examples: []string{
@@ -70,9 +72,11 @@ var buildinDialects = gherkinDialectMap{
 				"Կոնտեքստ",
 			},
 			scenario: []string{
+				"Օրինակ",
 				"Սցենար",
 			},
 			scenarioOutline: []string{
+				"Օրինակ կառուցվացքը",
 				"Սցենարի կառուցվացքը",
 			},
 			examples: []string{
@@ -110,9 +114,11 @@ var buildinDialects = gherkinDialectMap{
 				"Antecedents",
 			},
 			scenario: []string{
+				"Eixemplo",
 				"Caso",
 			},
 			scenarioOutline: []string{
+				"Esquema del eixemplo",
 				"Esquema del caso",
 			},
 			examples: []string{
@@ -155,6 +161,7 @@ var buildinDialects = gherkinDialectMap{
 				"الخلفية",
 			},
 			scenario: []string{
+				"مثال",
 				"سيناريو",
 			},
 			scenarioOutline: []string{
@@ -196,9 +203,11 @@ var buildinDialects = gherkinDialectMap{
 				"Antecedentes",
 			},
 			scenario: []string{
+				"Exemplo",
 				"Casu",
 			},
 			scenarioOutline: []string{
+				"Esbozu del exemplo",
 				"Esbozu del casu",
 			},
 			examples: []string{
@@ -240,6 +249,7 @@ var buildinDialects = gherkinDialectMap{
 				"Kontekst",
 			},
 			scenario: []string{
+				"Nümunələr",
 				"Ssenari",
 			},
 			scenarioOutline: []string{
@@ -283,9 +293,11 @@ var buildinDialects = gherkinDialectMap{
 				"Предистория",
 			},
 			scenario: []string{
+				"Пример",
 				"Сценарий",
 			},
 			scenarioOutline: []string{
+				"Рамка на пример",
 				"Рамка на сценарий",
 			},
 			examples: []string{
@@ -369,10 +381,12 @@ var buildinDialects = gherkinDialectMap{
 				"Pozadina",
 			},
 			scenario: []string{
+				"Primjer",
 				"Scenariju",
 				"Scenario",
 			},
 			scenarioOutline: []string{
+				"Primjer-obris",
 				"Scenariju-obris",
 				"Scenario-outline",
 			},
@@ -413,9 +427,11 @@ var buildinDialects = gherkinDialectMap{
 				"Antecedents",
 			},
 			scenario: []string{
+				"Exemple",
 				"Escenari",
 			},
 			scenarioOutline: []string{
+				"Esquema de l'exemple",
 				"Esquema de l'escenari",
 			},
 			examples: []string{
@@ -457,6 +473,7 @@ var buildinDialects = gherkinDialectMap{
 				"Kontext",
 			},
 			scenario: []string{
+				"Příklad",
 				"Scénář",
 			},
 			scenarioOutline: []string{
@@ -499,6 +516,7 @@ var buildinDialects = gherkinDialectMap{
 				"Cefndir",
 			},
 			scenario: []string{
+				"Enghraifft",
 				"Scenario",
 			},
 			scenarioOutline: []string{
@@ -538,9 +556,11 @@ var buildinDialects = gherkinDialectMap{
 				"Baggrund",
 			},
 			scenario: []string{
+				"Eksempel",
 				"Scenarie",
 			},
 			scenarioOutline: []string{
+				"Abstrakt Eksempel",
 				"Abstrakt Scenario",
 			},
 			examples: []string{
@@ -577,9 +597,11 @@ var buildinDialects = gherkinDialectMap{
 				"Grundlage",
 			},
 			scenario: []string{
+				"Beispiel",
 				"Szenario",
 			},
 			scenarioOutline: []string{
+				"Beispielrundriss",
 				"Szenariogrundriss",
 			},
 			examples: []string{
@@ -619,6 +641,7 @@ var buildinDialects = gherkinDialectMap{
 				"Υπόβαθρο",
 			},
 			scenario: []string{
+				"Παράδειγμα",
 				"Σενάριο",
 			},
 			scenarioOutline: []string{
@@ -660,6 +683,7 @@ var buildinDialects = gherkinDialectMap{
 				"💤",
 			},
 			scenario: []string{
+				"🥒",
 				"📕",
 			},
 			scenarioOutline: []string{
@@ -701,9 +725,11 @@ var buildinDialects = gherkinDialectMap{
 				"Background",
 			},
 			scenario: []string{
+				"Example",
 				"Scenario",
 			},
 			scenarioOutline: []string{
+				"Example Outline",
 				"Scenario Outline",
 				"Scenario Template",
 			},
@@ -955,10 +981,12 @@ var buildinDialects = gherkinDialectMap{
 				"Fono",
 			},
 			scenario: []string{
+				"Ekzemplo",
 				"Scenaro",
 				"Kazo",
 			},
 			scenarioOutline: []string{
+				"Konturo de la ekzemplo",
 				"Konturo de la scenaro",
 				"Skizo",
 				"Kazo-skizo",
@@ -998,9 +1026,11 @@ var buildinDialects = gherkinDialectMap{
 				"Antecedentes",
 			},
 			scenario: []string{
+				"Ejemplo",
 				"Escenario",
 			},
 			scenarioOutline: []string{
+				"Esquema del ejemplo",
 				"Esquema del escenario",
 			},
 			examples: []string{
@@ -1041,9 +1071,11 @@ var buildinDialects = gherkinDialectMap{
 				"Taust",
 			},
 			scenario: []string{
+				"Juhtum",
 				"Stsenaarium",
 			},
 			scenarioOutline: []string{
+				"Raamstjuhtum",
 				"Raamstsenaarium",
 			},
 			examples: []string{
@@ -1080,6 +1112,7 @@ var buildinDialects = gherkinDialectMap{
 				"زمینه",
 			},
 			scenario: []string{
+				"مثال",
 				"سناریو",
 			},
 			scenarioOutline: []string{
@@ -1158,9 +1191,11 @@ var buildinDialects = gherkinDialectMap{
 				"Contexte",
 			},
 			scenario: []string{
+				"Exemple",
 				"Scénario",
 			},
 			scenarioOutline: []string{
+				"Plan de l'exemple",
 				"Plan du scénario",
 				"Plan du Scénario",
 			},
@@ -1216,9 +1251,11 @@ var buildinDialects = gherkinDialectMap{
 				"Cúlra",
 			},
 			scenario: []string{
+				"Sampla",
 				"Cás",
 			},
 			scenarioOutline: []string{
+				"Sampla Achomair",
 				"Cás Achomair",
 			},
 			examples: []string{
@@ -1263,6 +1300,7 @@ var buildinDialects = gherkinDialectMap{
 				"બેકગ્રાઉન્ડ",
 			},
 			scenario: []string{
+				"ઉદાહરણ",
 				"સ્થિતિ",
 			},
 			scenarioOutline: []string{
@@ -1303,9 +1341,11 @@ var buildinDialects = gherkinDialectMap{
 				"Contexto",
 			},
 			scenario: []string{
+				"Exemplo",
 				"Escenario",
 			},
 			scenarioOutline: []string{
+				"Esbozo do exemplo",
 				"Esbozo do escenario",
 			},
 			examples: []string{
@@ -1347,6 +1387,7 @@ var buildinDialects = gherkinDialectMap{
 				"רקע",
 			},
 			scenario: []string{
+				"דוגמא",
 				"תרחיש",
 			},
 			scenarioOutline: []string{
@@ -1435,6 +1476,7 @@ var buildinDialects = gherkinDialectMap{
 				"Pozadina",
 			},
 			scenario: []string{
+				"Primjer",
 				"Scenarij",
 			},
 			scenarioOutline: []string{
@@ -1532,6 +1574,7 @@ var buildinDialects = gherkinDialectMap{
 				"Háttér",
 			},
 			scenario: []string{
+				"Példa",
 				"Forgatókönyv",
 			},
 			scenarioOutline: []string{
@@ -1654,9 +1697,11 @@ var buildinDialects = gherkinDialectMap{
 				"Contesto",
 			},
 			scenario: []string{
+				"Esempio",
 				"Scenario",
 			},
 			scenarioOutline: []string{
+				"Schema dello esempio",
 				"Schema dello scenario",
 			},
 			examples: []string{
@@ -1787,9 +1832,11 @@ var buildinDialects = gherkinDialectMap{
 				"კონტექსტი",
 			},
 			scenario: []string{
+				"მაგალითად",
 				"სცენარის",
 			},
 			scenarioOutline: []string{
+				"მაგალითად ნიმუში",
 				"სცენარის ნიმუში",
 			},
 			examples: []string{
@@ -1826,6 +1873,7 @@ var buildinDialects = gherkinDialectMap{
 				"ಹಿನ್ನೆಲೆ",
 			},
 			scenario: []string{
+				"ಉದಾಹರಣೆ",
 				"ಕಥಾಸಾರಾಂಶ",
 			},
 			scenarioOutline: []string{
@@ -1907,6 +1955,7 @@ var buildinDialects = gherkinDialectMap{
 				"Kontekstas",
 			},
 			scenario: []string{
+				"Pavyzdys",
 				"Scenarijus",
 			},
 			scenarioOutline: []string{
@@ -1948,9 +1997,11 @@ var buildinDialects = gherkinDialectMap{
 				"Hannergrond",
 			},
 			scenario: []string{
+				"Beispill",
 				"Szenario",
 			},
 			scenarioOutline: []string{
+				"Plang vum Beispill",
 				"Plang vum Szenario",
 			},
 			examples: []string{
@@ -1991,9 +2042,11 @@ var buildinDialects = gherkinDialectMap{
 				"Situācija",
 			},
 			scenario: []string{
+				"Piemērs",
 				"Scenārijs",
 			},
 			scenarioOutline: []string{
+				"Piemērs pēc parauga",
 				"Scenārijs pēc parauga",
 			},
 			examples: []string{
@@ -2034,6 +2087,7 @@ var buildinDialects = gherkinDialectMap{
 				"Содржина",
 			},
 			scenario: []string{
+				"Пример",
 				"Сценарио",
 				"На пример",
 			},
@@ -2169,9 +2223,11 @@ var buildinDialects = gherkinDialectMap{
 				"Achtergrond",
 			},
 			scenario: []string{
+				"Voorbeeld",
 				"Scenario",
 			},
 			scenarioOutline: []string{
+				"Abstract Voorbeeld",
 				"Abstract Scenario",
 			},
 			examples: []string{
@@ -2210,9 +2266,12 @@ var buildinDialects = gherkinDialectMap{
 				"Bakgrunn",
 			},
 			scenario: []string{
+				"Eksempel",
 				"Scenario",
 			},
 			scenarioOutline: []string{
+				"Eksempelmal",
+				"Abstrakt Eksempel",
 				"Scenariomal",
 				"Abstrakt Scenario",
 			},
@@ -2252,6 +2311,7 @@ var buildinDialects = gherkinDialectMap{
 				"ਪਿਛੋਕੜ",
 			},
 			scenario: []string{
+				"ਉਦਾਹਰਨ",
 				"ਪਟਕਥਾ",
 			},
 			scenarioOutline: []string{
@@ -2296,6 +2356,7 @@ var buildinDialects = gherkinDialectMap{
 				"Założenia",
 			},
 			scenario: []string{
+				"Przykład",
 				"Scenariusz",
 			},
 			scenarioOutline: []string{
@@ -2346,10 +2407,13 @@ var buildinDialects = gherkinDialectMap{
 				"Fundo",
 			},
 			scenario: []string{
+				"Exemplo",
 				"Cenário",
 				"Cenario",
 			},
 			scenarioOutline: []string{
+				"Esquema do Exemplo",
+				"Delineação do Exemplo",
 				"Esquema do Cenário",
 				"Esquema do Cenario",
 				"Delineação do Cenário",
@@ -2397,9 +2461,12 @@ var buildinDialects = gherkinDialectMap{
 				"Context",
 			},
 			scenario: []string{
+				"Exemplu",
 				"Scenariu",
 			},
 			scenarioOutline: []string{
+				"Structura exemplu",
+				"Structură exemplu",
 				"Structura scenariu",
 				"Structură scenariu",
 			},
@@ -2449,6 +2516,7 @@ var buildinDialects = gherkinDialectMap{
 				"Контекст",
 			},
 			scenario: []string{
+				"Пример",
 				"Сценарий",
 			},
 			scenarioOutline: []string{
@@ -2499,6 +2567,7 @@ var buildinDialects = gherkinDialectMap{
 				"Pozadie",
 			},
 			scenario: []string{
+				"Príklad",
 				"Scenár",
 			},
 			scenarioOutline: []string{
@@ -2553,8 +2622,8 @@ var buildinDialects = gherkinDialectMap{
 				"Ozadje",
 			},
 			scenario: []string{
-				"Scenarij",
 				"Primer",
+				"Scenarij",
 			},
 			scenarioOutline: []string{
 				"Struktura scenarija",
@@ -2608,6 +2677,7 @@ var buildinDialects = gherkinDialectMap{
 				"Позадина",
 			},
 			scenario: []string{
+				"Пример",
 				"Сценарио",
 				"Пример",
 			},
@@ -2747,6 +2817,7 @@ var buildinDialects = gherkinDialectMap{
 				"பின்னணி",
 			},
 			scenario: []string{
+				"உதாரணமாக",
 				"காட்சி",
 			},
 			scenarioOutline: []string{
@@ -2756,7 +2827,7 @@ var buildinDialects = gherkinDialectMap{
 			examples: []string{
 				"எடுத்துக்காட்டுகள்",
 				"காட்சிகள்",
-				" நிலைமைகளில்",
+				"நிலைமைகளில்",
 			},
 			given: []string{
 				"* ",
@@ -2833,6 +2904,7 @@ var buildinDialects = gherkinDialectMap{
 				"నేపథ్యం",
 			},
 			scenario: []string{
+				"ఉదాహరణ",
 				"సన్నివేశం",
 			},
 			scenarioOutline: []string{
@@ -2919,9 +2991,11 @@ var buildinDialects = gherkinDialectMap{
 				"Geçmiş",
 			},
 			scenario: []string{
+				"Örnek",
 				"Senaryo",
 			},
 			scenarioOutline: []string{
+				"Örnek taslağı",
 				"Senaryo taslağı",
 			},
 			examples: []string{
@@ -3002,6 +3076,7 @@ var buildinDialects = gherkinDialectMap{
 				"Передумова",
 			},
 			scenario: []string{
+				"Приклад",
 				"Сценарій",
 			},
 			scenarioOutline: []string{